Story chatbot/requirements.txt
streamlit==1.28.1
groq==0.4.1
google-generativeai==0.3.2
openai==1.3.5
requests==2.31.0
python-dotenv==1.0.0
gtts==2.4.0
Pillow==10.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
flask==2.3.2
